+++ backend/src/main/scala/com/softwaremill/bootzooka/Main.scala
@@ -6,7 +6,7 @@ import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.Http.ServerBinding
 import akka.stream.ActorMaterializer
-import com.softwaremill.bootzooka.user.Session
+import com.softwaremill.bootzooka.user.application.Session
 import com.softwaremill.session.{SessionConfig, SessionManager}
 import com.typesafe.scalalogging.StrictLogging
