+++ backend/src/main/scala/com/softwaremill/bootzooka/Routes.scala
@@ -1,9 +1,9 @@
 package com.softwaremill.bootzooka
 
 import akka.http.scaladsl.server.Directives._
-import com.softwaremill.bootzooka.utils.http.RoutesRequestWrapper
+import com.softwaremill.bootzooka.common.api.RoutesRequestWrapper
 import com.softwaremill.bootzooka.passwordreset.PasswordResetRoutes
-import com.softwaremill.bootzooka.user.UsersRoutes
+import com.softwaremill.bootzooka.user.api.UsersRoutes
 import com.softwaremill.bootzooka.version.VersionRoutes
 
 trait Routes extends RoutesRequestWrapper
