+++ backend/src/main/scala/com/softwaremill/bootzooka/common/sql/DatabaseConfig.scala
@@ -1,7 +1,15 @@
-package com.softwaremill.bootzooka.sql
+/*
+ * COPYRIGHT (c) 2016 VOCADO, LLC.  ALL RIGHTS RESERVED.  THIS SOFTWARE CONTAINS
+ * TRADE SECRETS AND/OR CONFIDENTIAL INFORMATION PROPRIETARY TO VOCADO, LLC AND/OR
+ * ITS LICENSORS. ACCESS TO AND USE OF THIS INFORMATION IS STRICTLY LIMITED AND
+ * CONTROLLED BY VOCADO, LLC.  THIS SOFTWARE MAY NOT BE COPIED, MODIFIED, DISTRIBUTED,
+ * DISPLAYED, DISCLOSED OR USED IN ANY WAY NOT EXPRESSLY AUTHORIZED BY VOCADO, LLC IN WRITING.
+ */
 
-import com.softwaremill.bootzooka.utils.ConfigWithDefault
-import com.softwaremill.bootzooka.sql.DatabaseConfig._
+package com.softwaremill.bootzooka.common.sql
+
+import com.softwaremill.bootzooka.common.ConfigWithDefault
+import com.softwaremill.bootzooka.common.sql.DatabaseConfig._
 import com.typesafe.config.Config
 
 trait DatabaseConfig extends ConfigWithDefault {
