+++ backend/src/main/scala/com/softwaremill/bootzooka/common/sql/SqlDatabase.scala
@@ -1,8 +1,17 @@
-package com.softwaremill.bootzooka.sql
+/*
+ * COPYRIGHT (c) 2016 VOCADO, LLC.  ALL RIGHTS RESERVED.  THIS SOFTWARE CONTAINS
+ * TRADE SECRETS AND/OR CONFIDENTIAL INFORMATION PROPRIETARY TO VOCADO, LLC AND/OR
+ * ITS LICENSORS. ACCESS TO AND USE OF THIS INFORMATION IS STRICTLY LIMITED AND
+ * CONTROLLED BY VOCADO, LLC.  THIS SOFTWARE MAY NOT BE COPIED, MODIFIED, DISTRIBUTED,
+ * DISPLAYED, DISCLOSED OR USED IN ANY WAY NOT EXPRESSLY AUTHORIZED BY VOCADO, LLC IN WRITING.
+ */
+
+package com.softwaremill.bootzooka.common.sql
 
 import java.net.URI
-import java.time.{ZoneOffset, OffsetDateTime}
-import DatabaseConfig._
+import java.time.{OffsetDateTime, ZoneOffset}
+
+import com.softwaremill.bootzooka.common.sql.DatabaseConfig._
 import com.typesafe.config.ConfigValueFactory._
 import com.typesafe.config.{Config, ConfigFactory}
 import com.typesafe.scalalogging.StrictLogging
