+++ backend/src/main/scala/com/softwaremill/bootzooka/user/application/UserService.scala
@@ -1,10 +1,20 @@
-package com.softwaremill.bootzooka.user
+/*
+ * COPYRIGHT (c) 2016 VOCADO, LLC.  ALL RIGHTS RESERVED.  THIS SOFTWARE CONTAINS
+ * TRADE SECRETS AND/OR CONFIDENTIAL INFORMATION PROPRIETARY TO VOCADO, LLC AND/OR
+ * ITS LICENSORS. ACCESS TO AND USE OF THIS INFORMATION IS STRICTLY LIMITED AND
+ * CONTROLLED BY VOCADO, LLC.  THIS SOFTWARE MAY NOT BE COPIED, MODIFIED, DISTRIBUTED,
+ * DISPLAYED, DISCLOSED OR USED IN ANY WAY NOT EXPRESSLY AUTHORIZED BY VOCADO, LLC IN WRITING.
+ */
 
-import java.time.{ZoneOffset, Instant}
+package com.softwaremill.bootzooka.user.application
+
+import java.time.{Instant, ZoneOffset}
 import java.util.UUID
 
-import com.softwaremill.bootzooka.utils.Utils
-import com.softwaremill.bootzooka.email.{EmailTemplatingEngine, EmailService}
+import com.softwaremill.bootzooka.common.Utils
+import com.softwaremill.bootzooka.email.{EmailService, EmailTemplatingEngine}
+import com.softwaremill.bootzooka.user._
+import com.softwaremill.bootzooka.user.domain.{BasicUserData, User}
 
 import scala.concurrent.{ExecutionContext, Future}
