--- backend/src/main/scala/com/softwaremill/bootzooka/user/rememberme/RememberMeToken.scala
@@ -1,6 +1,0 @@
-package com.softwaremill.bootzooka.user.rememberme
-
-import java.time.OffsetDateTime
-import java.util.UUID
-
-case class RememberMeToken(id: UUID, selector: String, tokenHash: String, userId: UUID, validTo: OffsetDateTime)
