+++ backend/src/test/scala/com/softwaremill/bootzooka/test/TestHelpersWithDb.scala
@@ -1,9 +1,10 @@
 package com.softwaremill.bootzooka.test
 
+import com.softwaremill.bootzooka.common.sql.SqlDatabase
 import com.softwaremill.bootzooka.config.CoreConfig
 import com.softwaremill.bootzooka.email.{DummyEmailService, EmailTemplatingEngine}
-import com.softwaremill.bootzooka.sql.SqlDatabase
-import com.softwaremill.bootzooka.user.{User, UserDao, UserService}
+import com.softwaremill.bootzooka.user.domain.User
+import com.softwaremill.bootzooka.user.application.{UserDao, UserService}
 import com.typesafe.config.ConfigFactory
 import org.scalatest.concurrent.ScalaFutures
