+++ backend/src/test/scala/com/softwaremill/bootzooka/user/EncryptPasswordBenchmark.scala
@@ -1,6 +1,7 @@
 package com.softwaremill.bootzooka.user
 
-import com.softwaremill.bootzooka.utils.Utils
+import com.softwaremill.bootzooka.common.Utils
+import com.softwaremill.bootzooka.user.domain.User
 
 // Run this locally to determine the desired number of iterations in PBKDF2
 object EncryptPasswordBenchmark extends App {
