+++ backend/src/test/scala/com/softwaremill/bootzooka/user/UserSpec.scala
@@ -1,5 +1,6 @@
 package com.softwaremill.bootzooka.user
 
+import com.softwaremill.bootzooka.user.domain.User
 import org.scalatest.{FlatSpec, Matchers}
 
 class UserSpec extends FlatSpec with Matchers {
