+++ ui-tests/src/test/scala/uitest/BaseUiSpec.scala
@@ -5,12 +5,12 @@ import java.util.concurrent.TimeUnit
 import akka.http.scaladsl.Http.ServerBinding
 import com.softwaremill.bootzooka.email.DummyEmailService
 import com.softwaremill.bootzooka.passwordreset.SqlPasswordResetCodeSchema
-import com.softwaremill.bootzooka.user.SqlUserSchema
+import com.softwaremill.bootzooka.user.application.SqlUserSchema
 import com.softwaremill.bootzooka.{BusinessLogic, Main}
 import org.openqa.selenium.firefox.FirefoxDriver
 import org.openqa.selenium.support.PageFactory
 import org.scalatest.concurrent.ScalaFutures
-import org.scalatest.{Matchers, BeforeAndAfter, BeforeAndAfterAll, FunSuite}
+import org.scalatest.{BeforeAndAfter, BeforeAndAfterAll, FunSuite, Matchers}
 import uitest.pages.{LoginPage, MainPage, MessagesPage, PasswordResetPage}
 
 import scala.util.Try
